+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Project name: quality-report-generator to reporter
 - Bugfix: Exporter Template
+- Bugfix: Patients pro Attribute
 
 ### Fixed
 - Intellij Warnings