+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2023-09-21]
+## [1.0.0 - 2023-09-22]
 ### Added
 - Gitignore
 - Dockerignore
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Clear Logger Buffer
 - Sort Source Paths in Context
 - Manage Exception during report generation
+- Custom Template ID
 
 
 ### Changed